ros and ubuntu setup

+++ setup_ubuntu.sh
@@ -1,46 +1,42 @@
 #!/bin/bash
 #Development environment setup in ubuntu
 # TODO
-# 1) check software installed before install them
-# 2) make printingout more clean
+# 1) make printingout more clean
 #
 
 #==================================================================
 #      			      FUNCTIONS
 
-# Detect the architecture
-if [[ "$(uname -m)" = "x86_64" ]]; then
-  ARCHITECTURE="x64"
-else
-  ARCHITECTURE="x32"
-fi
-
-function print_out()
-{
-	printf "\n$@\n"
-	#exit 2
-}
-
-
 
 #==================================================================
-#      			      MAIN FUNCTION
-print_out '------------------------------------------'
-print_out ' Install git, curl, terminator '
+#      			      SYSTEM TOOLS
+echo '----- SYSTEM TOOLS -------------------------'
 
+if [ -f '/usr/bin/curl' ]; then
+	echo "Curl installed";
+else
+	sudo apt-get install -y curl 
+fi
 
-#TOOLS
-sudo apt install -y curl
-
-sudo apt-get install -y terminator
+if [ -L '/usr/bin/terminator' ]; then
+	echo "Terminator installed";
+else
+	sudo apt-get install -y terminator 
+fi
 # write force_color_prompt=yes to .bashrc
 
 
-#Text editor (VIM + Sublime)
-sudo apt-get install -y vim
+#==================================================================
+#      			      PRODUCTIVITY
+echo '----- PRODUCTIVITY -------------------------'
 
-DIR='/opt/sublime_text'
-if [ -d $DIR ]; 
+if [ -L '/usr/bin/vim' ]; then
+	echo "Vim installed";
+else
+	sudo apt-get install -y vim 
+fi
+
+if [ -d '/opt/sublime_text' ]; 
   then 
 	echo "Sublime Text 3 installed";
   else
@@ -48,11 +44,25 @@ if [ -d $DIR ];
 	echo "deb https://download.sublimetext.com/ apt/stable/" | sudo tee /etc/apt/sources.list.d/sublime-text.list
 	sudo apt-get update
 	sudo apt-get install sublime-text
-	echo "Sublime Text 3 installed";
 #install sublime https://gist.github.com/simonewebdesign/8507139
 fi
 
-sudo apt autoremove -y
+if [ -f '/usr/bin/mark-my-words' ]; then
+	echo "Mark-my-words installed";
+else
+	echo "mark-my-words not installed";
+	sudo add-apt-repository ppa:voldyman/markmywords
+	sudo apt-get update
+	sudo apt-get install -y mark-my-words
+fi
+
+
+
+
+
+
+
+# sudo apt autoremove -y
 
 
 #==================================================================
@@ -60,4 +70,18 @@ sudo apt autoremove -y
 
 #touch exists.file
 #if [ -f exists.file ] ; then echo "yes" ; else echo "no" ; fi
+
+
+# # Detect the architecture
+# if [[ "$(uname -m)" = "x86_64" ]]; then
+#   ARCHITECTURE="x64"
+# else
+#   ARCHITECTURE="x32"
+# fi
+
+# function print_out()
+# {
+# 	printf "\n$@\n"
+# 	#exit 2
+# }
